UpdateFunctionality
Added update functionality
Moved getapi call to parent component for easier reach
